+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## v0.5.0 (unreleased)
 
-* cli: Allow to repeat the same chaos event with `-repeat` in order to terminate
-  multiple EC2 instances of an auto scaling group. Use `-repeat-interval` to
-  configure the time to wait before triggering the next event.
+* cli: Allow to repeat the same chaos event with `-count` in order to terminate
+  multiple EC2 instances of an auto scaling group. Use `-interval` to configure
+  the time to wait before triggering the next event.
 * cli: Set timeout of 10 seconds on AWS operations.
 * Build project with Go 1.7.